refactor(configuration): :art: Cleaned up corsOptions and some minor formatting changes
