+++ .editorconfig
@@ -1,7 +1,6 @@
 root=true
 
-[*]
-
+[*.{js,json,ts}]
 charset=utf-8
 end_of_line=lf
 insert_final_newline=true
@@ -9,3 +8,6 @@ insert_final_newline=true
 indent_style=space
 indent_size=2
 trim_trailing_whitespace=true
+
+[*.md]
+trim_trailing_whitespace=false
