+++ .eslintrc.json
@@ -8,6 +8,10 @@
     "ecmaVersion": 2018,
     "sourceType": "module"
   },
+  "env": {
+    "node": true,
+    "mocha": true
+  },
   "rules": {
     "@typescript-eslint/explicit-member-accessibility": 0,
     "@typescript-eslint/explicit-function-return-type": 0,
