+++ .gitignore
@@ -69,8 +69,8 @@ typings/
 .yarn-integrity
 
 # dotenv environment variables file
-.env
-.env.test
+.env*
+!.env*.example
 
 # parcel-bundler cache (https://parceljs.org/)
 .cache
