+++ .vscode/extensions.json
@@ -18,5 +18,6 @@
     "skyran.js-jsx-snippets",
     "redhat.vscode-yaml",
     "editorconfig.editorconfig"
-  ]
+  ],
+  "unwantedRecommendations": []
 }
