+++ tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "strict": true,
     "target": "ES2017",
     "lib": ["ES2017", "ESNext.AsyncIterable"],
     "typeRoots": ["./node_modules/@types", "src/@types"],
@@ -13,10 +14,13 @@
     "allowSyntheticDefaultImports": true,
     "forceConsistentCasingInFileNames": true,
     "outDir": "./build",
-    "baseUrl": "./src",
+    "baseUrl": "src",
     "module": "commonjs",
     "moduleResolution": "node",
-    "types": ["node"]
+    // "types": ["node"],
+    "paths": {
+      "config/*": ["config/*"]
+    }
   },
   "include": ["./src/**/*"],
   "exclude": ["node_modules"]
